+++ src/mac/LoRaMacHelper.cpp
@@ -36,6 +36,7 @@ LoRaMacRegion_t region;
 bool _otaa = false;
 
 bool _dutyCycleEnabled = false;
+extern bool PublicNetwork;
 
 bool lmh_mac_is_busy = false;
 
@@ -691,7 +692,7 @@ static void MlmeConfirm(MlmeConfirm_t *mlmeConfirm)
 }
 
 lmh_error_status lmh_init(lmh_callback_t *callbacks, lmh_param_t lora_param, bool otaa,
-						  eDeviceClass nodeClass, LoRaMacRegion_t user_region)
+						  eDeviceClass nodeClass, LoRaMacRegion_t user_region, bool region_change)
 {
 	region = (LoRaMacRegion_t)user_region;
 	char strlog1[64];
@@ -705,6 +706,8 @@ lmh_error_status lmh_init(lmh_callback_t *callbacks, lmh_param_t lora_param, boo
 	_otaa = otaa;
 
 	_dutyCycleEnabled = m_param.duty_cycle;
+
+	PublicNetwork = m_param.enable_public_network;
 
 #if (STATIC_DEVICE_EUI != 1)
 	m_callbacks->BoardGetUniqueId(DevEui);
@@ -742,7 +745,7 @@ lmh_error_status lmh_init(lmh_callback_t *callbacks, lmh_param_t lora_param, boo
 	LoRaMacPrimitives.MacMlmeConfirm = MlmeConfirm;
 	LoRaMacCallbacks.GetBatteryLevel = m_callbacks->BoardGetBatteryLevel;
 
-	error_status = LoRaMacInitialization(&LoRaMacPrimitives, &LoRaMacCallbacks, region, nodeClass);
+	error_status = LoRaMacInitialization(&LoRaMacPrimitives, &LoRaMacCallbacks, region, nodeClass, region_change);
 	if (error_status != LORAMAC_STATUS_OK)
 	{
 		return LMH_ERROR;